Straight removal of version_client and tests. Moved some used functions.
