+++ acceptancetests/jujupy/client.py
@@ -3442,3 +3442,65 @@ class GroupReporter:
         self.last_group = group
         self.ticks = 0
         self.wrap_offset = lead_length if lead_length < self.wrap_width else 0
+
+
+def _get_full_path(juju_path):
+    """Helper to ensure a full path is used.
+
+    If juju_path is None, ModelClient.get_full_path is used.  Otherwise,
+    the supplied path is converted to absolute.
+    """
+    if juju_path is None:
+        return ModelClient.get_full_path()
+    else:
+        return os.path.abspath(juju_path)
+
+
+def client_from_config(config, juju_path, debug=False, soft_deadline=None):
+    """Create a client from an environment's configuration.
+
+    :param config: Name of the environment to use the config from.
+    :param juju_path: Path to juju binary the client should wrap.
+    :param debug=False: The debug flag for the client, False by default.
+    :param soft_deadline: A datetime representing the deadline by which
+        normal operations should complete.  If None, no deadline is
+        enforced.
+    """
+    version = ModelClient.get_version(juju_path)
+    if config is None:
+        env = ModelClient.config_class('', {})
+    else:
+        env = ModelClient.config_class.from_config(config)
+    full_path = _get_full_path(juju_path)
+    return ModelClient(
+        env, version, full_path, debug=debug, soft_deadline=soft_deadline)
+
+
+def client_for_existing(juju_path, juju_data_dir, debug=False,
+                        soft_deadline=None, controller_name=None,
+                        model_name=None):
+    """Create a client for an existing controller/model.
+
+    :param juju_path: Path to juju binary the client should wrap.
+    :param juju_data_dir: Path to the juju data directory referring the the
+        controller and model.
+    :param debug=False: The debug flag for the client, False by default.
+    :param soft_deadline: A datetime representing the deadline by which
+        normal operations should complete.  If None, no deadline is
+        enforced.
+    """
+    version = ModelClient.get_version(juju_path)
+    full_path = _get_full_path(juju_path)
+    backend = ModelClient.default_backend(
+        full_path, version, set(), debug=debug, soft_deadline=soft_deadline)
+    if controller_name is None:
+        current_controller = backend.get_active_controller(juju_data_dir)
+        controller_name = current_controller
+    if model_name is None:
+        current_model = backend.get_active_model(juju_data_dir)
+        model_name = current_model
+    config = ModelClient.config_class.for_existing(
+        juju_data_dir, controller_name, model_name)
+    return ModelClient(
+        config, version, full_path,
+        debug=debug, soft_deadline=soft_deadline, _backend=backend)
